# task1/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get all messages
      tags:
        - messages
      responses:
        '200':
          description: A list of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      summary: Create a new message
      tags:
        - messages
      requestBody:
        description: The message to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: The created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /messages/{id}:
    patch:
      summary: Patch message by id
      tags:
        - messages
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the message to patch
          schema:
            type: integer
      requestBody:
        description: The message to patch
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        "201":
          description:  The message is patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
    delete:
      summary: Delete message by id
      tags: 
        - messages
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the message to delete
          schema:
            type: integer
      requestBody:
        description: The message to delete
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'  
      responses:
        "201":
          description:  The message is deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'    
components:
  schemas:
    Message:
      type: object
      properties:
        id:
          type: integer
          format: uint # так как в БД айди это только положительные числа.
        message:
          type: string